♻️ 동아리 Usecase 경로 변경
