+++ src/main/kotlin/kr/flooding/backend/domain/club/controller/ClubController.kt
@@ -9,18 +9,18 @@ import kr.flooding.backend.domain.club.dto.web.response.FetchClubFilterResponse
 import kr.flooding.backend.domain.club.dto.web.response.FetchClubMyselfResponse
 import kr.flooding.backend.domain.club.dto.web.response.FetchClubResponse
 import kr.flooding.backend.domain.club.enums.ClubType
-import kr.flooding.backend.domain.club.usecase.CloseClubUsecase
-import kr.flooding.backend.domain.club.usecase.ConfirmClubInviteUsecase
-import kr.flooding.backend.domain.club.usecase.CreateClubUsecase
-import kr.flooding.backend.domain.club.usecase.FetchClubFilterUsecase
-import kr.flooding.backend.domain.club.usecase.FetchClubMyselfUsecase
-import kr.flooding.backend.domain.club.usecase.FetchClubUsecase
-import kr.flooding.backend.domain.club.usecase.InviteClubMemberUsecase
-import kr.flooding.backend.domain.club.usecase.OpenClubUsecase
-import kr.flooding.backend.domain.club.usecase.RemoveClubMemberUsecase
-import kr.flooding.backend.domain.club.usecase.RemoveClubUsecase
-import kr.flooding.backend.domain.club.usecase.UpdateClubUsecase
-import kr.flooding.backend.domain.club.usecase.WithdrawClubUsecase
+import kr.flooding.backend.domain.club.usecase.student.CloseClubUsecase
+import kr.flooding.backend.domain.club.usecase.student.ConfirmClubInviteUsecase
+import kr.flooding.backend.domain.club.usecase.student.CreateClubUsecase
+import kr.flooding.backend.domain.club.usecase.student.FetchClubFilterUsecase
+import kr.flooding.backend.domain.club.usecase.student.FetchClubMyselfUsecase
+import kr.flooding.backend.domain.club.usecase.student.FetchClubUsecase
+import kr.flooding.backend.domain.club.usecase.student.InviteClubMemberUsecase
+import kr.flooding.backend.domain.club.usecase.student.OpenClubUsecase
+import kr.flooding.backend.domain.club.usecase.student.RemoveClubMemberUsecase
+import kr.flooding.backend.domain.club.usecase.student.RemoveClubUsecase
+import kr.flooding.backend.domain.club.usecase.student.UpdateClubUsecase
+import kr.flooding.backend.domain.club.usecase.student.WithdrawClubUsecase
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
