+++ src/main/kotlin/kr/flooding/backend/domain/club/usecase/student/OpenClubUsecase.kt
@@ -1,32 +1,36 @@
-package kr.flooding.backend.domain.club.usecase
+package kr.flooding.backend.domain.club.usecase.student
 
-import jakarta.transaction.Transactional
+import kr.flooding.backend.domain.club.enums.ClubStatus
 import kr.flooding.backend.domain.club.persistence.repository.ClubRepository
 import kr.flooding.backend.global.exception.ExceptionEnum
 import kr.flooding.backend.global.exception.HttpException
 import kr.flooding.backend.global.exception.toPair
 import kr.flooding.backend.global.util.UserUtil
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 import java.util.UUID
 
 @Service
 @Transactional
-class RemoveClubUsecase(
-	val clubRepository: ClubRepository,
-	val userUtil: UserUtil,
+class OpenClubUsecase(
+	private val userUtil: UserUtil,
+	private val clubRepository: ClubRepository,
 ) {
 	fun execute(clubId: UUID) {
-		val user = userUtil.getUser()
-
+		val currentUser = userUtil.getUser()
 		val club =
 			clubRepository.findById(clubId).orElseThrow {
-				throw HttpException(ExceptionEnum.CLUB.NOT_FOUND_CLUB.toPair())
+				HttpException(ExceptionEnum.CLUB.NOT_FOUND_CLUB.toPair())
 			}
 
-		if (user != club.leader) {
+		if (currentUser != club.leader) {
 			throw HttpException(ExceptionEnum.CLUB.NOT_CLUB_LEADER.toPair())
 		}
 
-		clubRepository.delete(club)
+		if (club.status != ClubStatus.APPROVED) {
+			throw HttpException(ExceptionEnum.CLUB.NOT_APPROVED_CLUB.toPair())
+		}
+
+		club.startRecruitment()
 	}
 }
