+++ src/main/kotlin/kr/flooding/backend/domain/club/usecase/student/RemoveClubUsecase.kt
@@ -1,36 +1,32 @@
-package kr.flooding.backend.domain.club.usecase
+package kr.flooding.backend.domain.club.usecase.student
 
-import kr.flooding.backend.domain.club.enums.ClubStatus
+import jakarta.transaction.Transactional
 import kr.flooding.backend.domain.club.persistence.repository.ClubRepository
 import kr.flooding.backend.global.exception.ExceptionEnum
 import kr.flooding.backend.global.exception.HttpException
 import kr.flooding.backend.global.exception.toPair
 import kr.flooding.backend.global.util.UserUtil
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
 import java.util.UUID
 
 @Service
 @Transactional
-class CloseClubUsecase(
-	private val userUtil: UserUtil,
-	private val clubRepository: ClubRepository,
+class RemoveClubUsecase(
+	val clubRepository: ClubRepository,
+	val userUtil: UserUtil,
 ) {
 	fun execute(clubId: UUID) {
-		val currentUser = userUtil.getUser()
+		val user = userUtil.getUser()
+
 		val club =
 			clubRepository.findById(clubId).orElseThrow {
-				HttpException(ExceptionEnum.CLUB.NOT_FOUND_CLUB.toPair())
+				throw HttpException(ExceptionEnum.CLUB.NOT_FOUND_CLUB.toPair())
 			}
 
-		if (currentUser != club.leader) {
+		if (user != club.leader) {
 			throw HttpException(ExceptionEnum.CLUB.NOT_CLUB_LEADER.toPair())
 		}
 
-		if (club.status != ClubStatus.APPROVED) {
-			throw HttpException(ExceptionEnum.CLUB.NOT_APPROVED_CLUB.toPair())
-		}
-
-		club.stopRecruitment()
+		clubRepository.delete(club)
 	}
 }
