+++ src/main/kotlin/kr/flooding/backend/domain/club/usecase/student/UpdateClubUsecase.kt
@@ -1,4 +1,4 @@
-package kr.flooding.backend.domain.club.usecase
+package kr.flooding.backend.domain.club.usecase.student
 
 import kr.flooding.backend.domain.club.dto.web.request.UpdateClubRequest
 import kr.flooding.backend.domain.club.persistence.repository.ClubRepository
